refactor: merge events service into metrics service
- Move external_log_collector functionality to metrics/rest/events controller module
- Add 'origin' attribute to EventBase to track event source (e.g., 'kdcube-frontend')
- Mount events router in metrics service alongside existing metrics endpoints
- Update nginx configs to route /api/events/* to metrics service (port 8090)
- Remove separate chat-frontend-logs-collector service from docker-compose
- Update Dockerfile_Collector to run metrics service instead

This consolidates event collection into the existing metrics service for faster integration of client event propagation feature.
