+++ app/ai-app/deployment/docker/all_in_one_kdcube/nginx/conf/nginx_proxy.conf
@@ -53,7 +53,7 @@ http {
     }
 
     upstream events_collector {
-        server chat-frontend-logs-collector:8080;
+        server metrics:8090;
     }
 
     upstream monitoring_api {
