+++ app/ai-app/deployment/docker/custom-ui-managed-infra/Dockerfile_Collector
@@ -1,4 +1,4 @@
-# Multi-stage build for external log collector service
+# Multi-stage build for metrics service (includes events collection)
 
 # Stage 1: Build stage
 FROM python:3.12-slim as builder
@@ -47,4 +47,4 @@ COPY --chown=appuser:appuser services/kdcube-ai-app/ .
 RUN chown -R appuser:appuser /app
 
 USER appuser
-CMD ["python", "kdcube_ai_app/apps/chat/external_log_collector/web_app.py"]
+CMD ["python", "kdcube_ai_app/apps/metrics/web_app.py"]
