+++ app/ai-app/deployment/docker/custom-ui-managed-infra/nginx/conf/nginx_proxy_delegated.conf
@@ -62,7 +62,7 @@ http {
     }
 
     upstream events_collector {
-        server chat-frontend-logs-collector:8080;
+        server metrics:8090;
     }
 
     upstream monitoring_api {
