+++ app/ai-app/services/kdcube-ai-app/kdcube_ai_app/apps/metrics/rest/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2026 Elena Viter
