+++ app/ai-app/services/kdcube-ai-app/kdcube_ai_app/apps/metrics/rest/events/event_type.py
@@ -4,6 +4,7 @@ from datetime import datetime
 
 class EventBase(BaseModel):
     event_type : Literal["log", "metric"]
+    origin: str  # Denotes the origin of the event (e.g., 'kdcube-frontend')
 
     tenant: str
     project: str
